+++ trunk/puzzle/Puzzle Core/src/org/puzzle/puzzlecore/swing/toolbox/widgettool/svg2mif/element/Dialogue.java
@@ -358,8 +358,8 @@ private void processPath(Element eec) {
 	       				if (e.equals("M")) {
 	       					po.addPoint((DPoint)premierPoint.clone());
 	       					DPoint center = new DPoint(0, 0);
-	       					center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-	       					center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+	       					center.x = (bounds[0].x + bounds[1].x) / 2;
+	       					center.y = (bounds[0].y + bounds[1].y) / 2;
 	       					po.setCenter((DPoint)center.clone());
 	       					MM.addObject(po.clone());
 	       					bounds[0] = new DPoint(Double.MAX_VALUE, Double.MAX_VALUE);
@@ -375,8 +375,8 @@ private void processPath(Element eec) {
 	       				if (e.equals("m")) {
 	       					po.addPoint((DPoint)premierPoint.clone());
 	       					DPoint center = new DPoint(0, 0);
-	       					center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-	       					center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+	       					center.x = (bounds[0].x + bounds[1].x) / 2;
+	       					center.y = (bounds[0].y + bounds[1].y) / 2;
 	       					po.setCenter((DPoint)center.clone());
 	       					MM.addObject(po.clone());
 	       					bounds[0] = new DPoint(Double.MAX_VALUE, Double.MAX_VALUE);
@@ -415,8 +415,8 @@ private void processPath(Element eec) {
 						}
 	       			}
    					DPoint center = new DPoint(0, 0);
-   					center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-   					center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+   					center.x = (bounds[0].x + bounds[1].x) / 2;
+   					center.y = (bounds[0].y + bounds[1].y) / 2;
    					po.setCenter((DPoint)center.clone());
    					po.addPoint((DPoint)premierPoint.clone());	       			
 	       			MM.addObject(po.clone());
@@ -460,8 +460,8 @@ private void processPath(Element eec) {
 			po.addPoint(new DPoint(x1, y2));
 			po.addPoint(new DPoint(x1, y1));
 			DPoint center = new DPoint(0, 0);
-			center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-			center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+			center.x = (bounds[0].x + bounds[1].x) / 2;
+			center.y = (bounds[0].y + bounds[1].y) / 2;
 			po.setCenter((DPoint)center.clone());
 			MM.addObject(po);
 		}
@@ -501,8 +501,8 @@ private void processPath(Element eec) {
 			po.addPoint(new DPoint(x1, y2));
 			po.addPoint(new DPoint(x1, y1));
 			DPoint center = new DPoint(0, 0);
-			center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-			center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+			center.x = (bounds[0].x + bounds[1].x) / 2;
+			center.y = (bounds[0].y + bounds[1].y) / 2;
 			po.setCenter((DPoint)center.clone());
 			MM.addObject(po);
 		}
@@ -553,8 +553,8 @@ private void processPath(Element eec) {
 				po.addPoint((DPoint) pt.clone());
 			}
 			DPoint center = new DPoint(0, 0);
-	        center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-			center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+                        center.x = (bounds[0].x + bounds[1].x) / 2;
+			center.y = (bounds[0].y + bounds[1].y) / 2;
 			po.setCenter((DPoint)center.clone());
 			po.addPoint((DPoint)premierPoint.clone());	  
 			MM.addObject(po);
@@ -582,8 +582,8 @@ private void processPath(Element eec) {
 			po.addPoint((DPoint) pt.clone());
 		}
 		DPoint center = new DPoint(0, 0);
-		center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-		center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+		center.x = (bounds[0].x + bounds[1].x) / 2;
+		center.y = (bounds[0].y + bounds[1].y) / 2;
 		po.setCenter(center);
 		MM.addObject(po);
 	}
@@ -628,8 +628,8 @@ private void processPath(Element eec) {
 			po.addPoint(new DPoint(x1, y2));
 			po.addPoint(new DPoint(x1, y1));
 			DPoint center = new DPoint(0, 0);
-			center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-			center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+			center.x = (bounds[0].x + bounds[1].x) / 2;
+			center.y = (bounds[0].y + bounds[1].y) / 2;
 			po.setCenter((DPoint)center.clone());
 			MM.addObject(po);
 		}
@@ -677,8 +677,8 @@ private void processPath(Element eec) {
 			po.addPoint(new DPoint(x1, y2));
 			po.addPoint(new DPoint(x1, y1));
 			DPoint center = new DPoint(0, 0);
-			center.x = (double) ((bounds[0].x + bounds[1].x) / 2);
-			center.y = (double) ((bounds[0].y + bounds[1].y) / 2);
+			center.x = (bounds[0].x + bounds[1].x) / 2;
+			center.y = (bounds[0].y + bounds[1].y) / 2;
 			po.setCenter((DPoint)center.clone());
 			MM.addObject(po);
 		}
